Initial commit (after Oct 1 2011 updates)
updates which were sent to AUR
and largely prompted by orbisvicis

I decided that these PKGBUILDs really deserve version control
AUR doesn't offer any currently
So I'm trying github - it may be valuable even if
no one takes advantage of the collaborative aspects.

Hopefully the risk of diverging (code on github and aur
and what if they are not both up-to-date; what if maintainers
change; etc.) is not bad enough to outweigh possible benefits.